Binary Converter class created
