+++ src/components/Pages/Dashboard/UserPicture/UserPicture.js
@@ -1,5 +1,6 @@
 import React from "react"
 import "./UserPicture.css"
+import BConvert from "../../../../utils/BinaryConverter"
 
 const UserPicture = ({user}) => {
 
@@ -10,31 +11,14 @@ const UserPicture = ({user}) => {
      **/
 
     // Raw binary array converted into an object
-    const imgObject = uintToString(user.image.image.data)
+    const imgObject = BConvert.uintToString(user.image.image.data)
 
     // Converting the image buffer data into a binary string
-    const bin = arrayBufferBase64(imgObject.buffer.data)
+    const bin = BConvert.arrayBufferBase64(imgObject.buffer.data)
 
     // In order to display as an image need the data:image prefix
-    // The image is ready to use as a corect image source
+    // The image is ready to use as a correct image source
     const imgSrc = `data:image/jpeg;base64,${bin}`
-
-    /** It converts binary array into base64 string **/
-    function arrayBufferBase64(buffer) {
-        let binary = ''
-        let bytes = [].slice.call(new Uint8Array(buffer))
-
-        bytes.forEach( b => binary += String.fromCharCode(b))
-
-        return btoa(binary)
-    }
-
-    /** Converts the binary data into an object **/
-    function uintToString(uintArray) {
-        let encodedString = String.fromCharCode.apply(null, uintArray),
-            decodedString = decodeURIComponent(escape(encodedString));
-        return JSON.parse(decodedString)
-    }
